Create faculty website using c++
This project aims to provide services for a site such as the college’s website.
This provides the student with the calculation of the percentage, absence rate, grade point average and the number of his lectures in one month. The doctor is allowed to calculate his salary and record doctors ’data and the number of his lectures in one month, and the project includes important points like Abstraction, Encapsulation, Polymorphism, Function Overloading, Function Overriding.